requirements.txt
llama-cpp-python
langchain
streamlit==1.31.0
pathlib
sentence-transformers
farm-haystack
farm-haystack[all]
ollama